+++ lib/simple_oauth/oauth2/error.rb
@@ -1,3 +1,4 @@
+require_relative "../errors"
 require_relative "response_body"
 
 module SimpleOAuth
@@ -7,7 +8,7 @@ module SimpleOAuth
     # @api public
     # @example Raise the error described by a failed response
     #   raise SimpleOAuth::OAuth2::Error.from_response(status: 400, body: '{"error":"invalid_grant"}')
-    class Error < StandardError
+    class Error < SimpleOAuth::Error
       # The error message for a status that is not an HTTP status
       INVALID_STATUS = "The status must be an Integer or a String of digits".freeze